[DOP-20962] Add scheduler integration test
